Refactor out static method
